Ellipsis menu in block navigator (#22427)
* Super early prototype

* Hide <BlockSettingsMenu /> in the top level item

* Hide the ellipsis menu behind __experimentalWithEllipsisMenu flag

* Move the menu to the same level as mover cells

* Use a vertical icon for the ellipsis menu

* Support for arbitrary clientIds in BlockSettingsMenuControlsSlot

* Use a smaller dropdown button in the block navigator

* Add the ellipsis menu to the post editor block navigation

* Reduce the size of movers

* Update packages/edit-navigation/src/components/menu-editor/navigation-structure-panel.js


* Remove unnecessary arg to useBlockNavigator

* Pass props to nested BlockSettingsDropdown
