+++ packages/block-editor/src/components/block-navigation/context.js
@@ -5,6 +5,8 @@ import { createContext, useContext } from '@wordpress/element';
 
 export const BlockNavigationContext = createContext( {
 	__experimentalWithBlockNavigationSlots: false,
+	__experimentalWithEllipsisMenu: false,
+	__experimentalWithEllipsisMenuMinLevel: 0,
 } );
 
 export const useBlockNavigationContext = () =>
