+++ packages/block-editor/src/components/block-navigation/tree.js
@@ -21,11 +21,24 @@ import { BlockNavigationContext } from './context';
  */
 export default function BlockNavigationTree( {
 	__experimentalWithBlockNavigationSlots,
+	__experimentalWithEllipsisMenu,
+	__experimentalWithEllipsisMenuMinLevel,
 	...props
 } ) {
 	const contextValue = useMemo(
-		() => ( { __experimentalWithBlockNavigationSlots } ),
-		[ __experimentalWithBlockNavigationSlots ]
+		() => ( {
+			__experimentalWithBlockNavigationSlots,
+			__experimentalWithEllipsisMenu,
+			__experimentalWithEllipsisMenuMinLevel:
+				typeof __experimentalWithEllipsisMenuMinLevel === 'number'
+					? __experimentalWithEllipsisMenuMinLevel
+					: 0,
+		} ),
+		[
+			__experimentalWithBlockNavigationSlots,
+			__experimentalWithEllipsisMenu,
+			__experimentalWithEllipsisMenuMinLevel,
+		]
 	);
 
 	return (
