+++ packages/edit-navigation/src/components/menu-editor/navigation-structure-panel.js
@@ -25,7 +25,9 @@ export default function NavigationStructurePanel( { blocks, initialOpen } ) {
 						blocks={ blocks }
 						selectedBlockClientId={ selectedBlockClientIds[ 0 ] }
 						selectBlock={ selectBlock }
-						__experimentalWithBlockNavigationSlots={ true }
+						__experimentalWithBlockNavigationSlots
+						__experimentalWithEllipsisMenu
+						__experimentalWithEllipsisMenuMinLevel={ 2 }
 						showNestedBlocks
 						showAppender
 						showBlockMovers
